Fix de 90 pourcent des problemes de la seance de test du 24/11/2021

+++ routes/web.php
@@ -146,43 +146,43 @@ Route::middleware(['auth','verified'])->namespace('Form')->group(function () {
     Route::get('forms', 'FormController@index')->name('forms.index');
     Route::get('forms/create', 'FormController@create')->name('forms.create');
     Route::post('forms', 'FormController@store')->name('forms.store');
-    Route::get('forms/{form}', 'FormController@show')->name('forms.show')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/conditional/{form}', 'FormController@conditional')->name('forms.conditional')->middleware('Role:Superadmin|Client|Free');
-    Route::post('forms/conditional', 'FormController@conditionalpost')->name('forms.conditionalpost')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/{form}/edit', 'FormController@edit')->name('forms.edit')->middleware('Role:Superadmin|Client|Free');
-    Route::put('forms/{form}', 'FormController@update')->name('forms.update')->middleware('Role:Superadmin|Client|Free');
+    Route::get('forms/{form}', 'FormController@show')->name('forms.show')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/conditional/{form}', 'FormController@conditional')->name('forms.conditional')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::post('forms/conditional', 'FormController@conditionalpost')->name('forms.conditionalpost')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/{form}/edit', 'FormController@edit')->name('forms.edit')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::put('forms/{form}', 'FormController@update')->name('forms.update')->middleware('Role:Superadmin|Client|Free|Admin');
     Route::delete('forms/{form}', 'FormController@destroy')->name('forms.destroy')->middleware('Role:Superadmin|Client');
 
-    Route::post('forms/{form}/draft', 'FormController@draftForm')->name('forms.draft')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/{form}/preview', 'FormController@previewForm')->name('forms.preview')->middleware('Role:Superadmin|Client|Free');
-    Route::post('forms/{form}/open', 'FormController@openFormForResponse')->name('forms.open')->middleware('Role:Superadmin|Client|Free');
-    Route::post('forms/{form}/close', 'FormController@closeFormToResponse')->name('forms.close')->middleware('Role:Superadmin|Client|Free');
+    Route::post('forms/{form}/draft', 'FormController@draftForm')->name('forms.draft')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/{form}/preview', 'FormController@previewForm')->name('forms.preview')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::post('forms/{form}/open', 'FormController@openFormForResponse')->name('forms.open')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::post('forms/{form}/close', 'FormController@closeFormToResponse')->name('forms.close')->middleware('Role:Superadmin|Client|Free|Admin');
 
-    Route::post('forms/{form}/share-via-email', 'FormController@shareViaEmail')->name('form.share.email')->middleware('Role:Superadmin|Client|Free');
-    Route::post('forms/{form}/form-availability', 'FormAvailabilityController@save')->name('form.availability.save')->middleware('Role:Superadmin|Client|Free');
-    Route::delete('forms/{form}/form-availability/reset', 'FormAvailabilityController@reset')->name('form.availability.reset')->middleware('Role:Superadmin|Client|Free');
+    Route::post('forms/{form}/share-via-email', 'FormController@shareViaEmail')->name('form.share.email')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::post('forms/{form}/form-availability', 'FormAvailabilityController@save')->name('form.availability.save')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::delete('forms/{form}/form-availability/reset', 'FormAvailabilityController@reset')->name('form.availability.reset')->middleware('Role:Superadmin|Client|Free|Admin');
 
     //Form Field Routes
-    Route::post('forms/{form}/fields/add', 'FieldController@store')->name('forms.fields.store')->middleware('Role:Superadmin|Client|Free');
-    Route::post('forms/{form}/fields/delete', 'FieldController@destroy')->name('forms.fields.destroy')->middleware('Role:Superadmin|Client|Free');
+    Route::post('forms/{form}/fields/add', 'FieldController@store')->name('forms.fields.store')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::post('forms/{form}/fields/delete', 'FieldController@destroy')->name('forms.fields.destroy')->middleware('Role:Superadmin|Client|Free|Admin');
 
     //Form Response Routes
-    Route::get('forms/{form}/responses', 'ResponseController@index')->name('forms.responses.index')->middleware('Role:Superadmin|Client|Free');
+    Route::get('forms/{form}/responses', 'ResponseController@index')->name('forms.responses.index')->middleware('Role:Superadmin|Client|Free|Admin');
 
-    Route::get('forms/{form}/responses/download', 'ResponseController@export')->name('forms.response.export')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/{form}/responses/download/{country_id}', 'ResponseController@exportcountry')->name('forms.response.exportcountry')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/{form}/responses/downloadsite/{site_id}', 'ResponseController@exportsite')->name('forms.response.exportsite')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/{form}/responses/downloadville/{ville}', 'ResponseController@exportville')->name('forms.response.exportville')->middleware('Role:Superadmin|Client|Free');
-    Route::get('forms/{form}/responses/downloaduser/{user_id}', 'ResponseController@exportuser')->name('forms.response.exportuser')->middleware('Role:Superadmin|Client|Free');
+    Route::get('forms/{form}/responses/download', 'ResponseController@export')->name('forms.response.export')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/{form}/responses/download/{country_id}', 'ResponseController@exportcountry')->name('forms.response.exportcountry')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/{form}/responses/downloadsite/{site_id}', 'ResponseController@exportsite')->name('forms.response.exportsite')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/{form}/responses/downloadville/{ville}', 'ResponseController@exportville')->name('forms.response.exportville')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::get('forms/{form}/responses/downloaduser/{user_id}', 'ResponseController@exportuser')->name('forms.response.exportuser')->middleware('Role:Superadmin|Client|Free|Admin');
 
 
 
-    Route::delete('forms/{form}/responses', 'ResponseController@destroyAll')->name('forms.responses.destroy.all')->middleware('Role:Superadmin|Client|Free');
-    Route::delete('forms/{form}/responses/{response}', 'ResponseController@destroy')->name('forms.responses.destroy.single')->middleware('Role:Superadmin|Client|Free');
+    Route::delete('forms/{form}/responses', 'ResponseController@destroyAll')->name('forms.responses.destroy.all')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::delete('forms/{form}/responses/{response}', 'ResponseController@destroy')->name('forms.responses.destroy.single')->middleware('Role:Superadmin|Client|Free|Admin');
 
     //Form Collaborator Routes
-    Route::post('forms/{form}/collaborators', 'CollaboratorController@store')->name('form.collaborators.store')->middleware('Role:Superadmin|Client|Free');
-    Route::delete('forms/{form}/collaborators/{collaborator}', 'CollaboratorController@destroy')->name('form.collaborator.destroy')->middleware('Role:Superadmin|Client|Free');
+    Route::post('forms/{form}/collaborators', 'CollaboratorController@store')->name('form.collaborators.store')->middleware('Role:Superadmin|Client|Free|Admin');
+    Route::delete('forms/{form}/collaborators/{collaborator}', 'CollaboratorController@destroy')->name('form.collaborator.destroy')->middleware('Role:Superadmin|Client|Free|Admin');
 });
